Moving efficiency processes to the task correlation
