+++ PWGHF/HFC/TableProducer/correlatorDsHadrons.cxx
@@ -153,8 +153,6 @@ struct HfCorrelatorDsHadrons {
   Configurable<int> selectionFlagDs{"selectionFlagDs", 7, "Selection Flag for Ds"};
   Configurable<int> numberEventsMixed{"numberEventsMixed", 5, "Number of events mixed in ME process"};
   Configurable<int> nTpcCrossedRaws{"nTpcCrossedRaws", 70, "Number of crossed TPC Rows"};
-  Configurable<int> eventGeneratorType{"eventGeneratorType", -1, "If positive, enable event selection using subGeneratorId information. The value indicates which events to keep (0 = MB, 4 = charm triggered, 5 = beauty triggered)"};
-  Configurable<bool> useSel8ForTrackEff{"useSel8ForTrackEff", true, "Flag for applying sel8 for collision selection"};
   Configurable<bool> applyEfficiency{"applyEfficiency", true, "Flag for applying D-meson efficiency weights"};
   Configurable<float> yCandMax{"yCandMax", 0.8, "max. cand. rapidity"};
   Configurable<float> yCandGenMax{"yCandGenMax", 0.5, "max. gen. cand. rapidity"};
@@ -163,7 +161,6 @@ struct HfCorrelatorDsHadrons {
   Configurable<float> dcaZTrackMax{"dcaZTrackMax", 1., "max. DCA_z of tracks"};
   Configurable<float> ptCandMin{"ptCandMin", 1., "min. cand. pT"};
   Configurable<float> ptCandMax{"ptCandMax", 50., "max. cand pT"};
-  Configurable<float> ptDaughterMin{"ptDaughterMin", 0.1, "min. daughter pT"};
   Configurable<float> ptTrackMin{"ptTrackMin", 0.3, "min. track pT"};
   Configurable<float> ptTrackMax{"ptTrackMax", 50., "max. track pT"};
   Configurable<std::vector<int>> classMl{"classMl", {0, 1, 2}, "Indexes of ML scores to be stored. Three indexes max."};
@@ -184,24 +181,6 @@ struct HfCorrelatorDsHadrons {
 
   HfHelper hfHelper;
   SliceCache cache;
-
-  enum CandidateStep { kCandidateStepMcGenAll = 0,
-                       kCandidateStepMcGenDsToKKPi,
-                       kCandidateStepMcCandInAcceptance,
-                       kCandidateStepMcDaughtersInAcceptance,
-                       kCandidateStepMcReco,
-                       kCandidateStepMcRecoInAcceptance,
-                       kCandidateNSteps };
-
-  enum AssocTrackStep { kAssocTrackStepMcGen = 0,
-                        kAssocTrackStepMcGenInAcceptance,
-                        kAssocTrackStepRecoAll,
-                        kAssocTrackStepRecoMcMatch,
-                        kAssocTrackStepRecoPrimaries,
-                        kAssocTrackStepRecoSpecies,
-                        kAssocTrackStepRecoCollMatch,
-                        kAssocTrackStepFake,
-                        kAssocTrackNSteps };
 
   using SelCollisionsWithDs = soa::Filtered<soa::Join<aod::Collisions, aod::Mults, aod::EvSels, aod::DmesonSelection>>;            // collisionFilter applied
   using SelCollisionsWithDsMc = soa::Filtered<soa::Join<aod::McCollisions, aod::DmesonSelection>>;                                 // collisionFilter applied
@@ -277,16 +256,6 @@ struct HfCorrelatorDsHadrons {
     registry.add("hPtParticleAssocMcGen", "Associated Particle - MC Gen", {HistType::kTH1F, {axisPtHadron}});
     registry.add("hEtaMcGen", "Ds,Hadron particles - MC Gen", {HistType::kTH1F, {axisEta}});
     registry.add("hPhiMcGen", "Ds,Hadron particles - MC Gen", {HistType::kTH1F, {axisPhi}});
-    // Histograms for efficiencies
-    auto hCandidates = registry.add<StepTHn>("hCandidates", "Candidate count at different steps", {HistType::kStepTHnF, {axisPtD, axisMultFT0M, {RecoDecay::OriginType::NonPrompt + 1, +RecoDecay::OriginType::None - 0.5, +RecoDecay::OriginType::NonPrompt + 0.5}}, kCandidateNSteps});
-    hCandidates->GetAxis(0)->SetTitle("#it{p}_{T} (GeV/#it{c})");
-    hCandidates->GetAxis(1)->SetTitle("multiplicity");
-    hCandidates->GetAxis(2)->SetTitle("Charm hadron origin");
-    auto hAssocTracks = registry.add<StepTHn>("hAssocTracks", "Associated tracks at different steps", {HistType::kStepTHnF, {axisEta, axisPtHadron, axisMultFT0M, axisPosZ}, kAssocTrackNSteps});
-    hAssocTracks->GetAxis(0)->SetTitle("#eta");
-    hAssocTracks->GetAxis(1)->SetTitle("#it{p}_{T} (GeV/#it{c})");
-    hAssocTracks->GetAxis(2)->SetTitle("multiplicity");
-    hAssocTracks->GetAxis(3)->SetTitle("pos z");
   }
 
   /// Fill histograms of quantities independent from the daugther-mass hypothesis for data
@@ -587,130 +556,6 @@ struct HfCorrelatorDsHadrons {
     } // end candidate loop
   }
   PROCESS_SWITCH(HfCorrelatorDsHadrons, processMcRec, "Process MC Reco mode", false);
-
-  /// Ds-Hadron correlation - for calculating candidate reconstruction efficiency using MC reco-level analysis
-  void processMcCandEfficiency(soa::Join<aod::Collisions, aod::FT0Mults> const&,
-                               soa::Join<aod::McCollisions, aod::MultsExtraMC> const&,
-                               CandDsMcGen const& mcParticles,
-                               CandDsMcReco const& candidates,
-                               aod::TracksWMc const&)
-  {
-    auto hCandidates = registry.get<StepTHn>(HIST("hCandidates"));
-
-    /// Gen loop
-    float multiplicity = -1.;
-    for (auto& mcParticle : mcParticles) {
-      // generated candidates
-      if (std::abs(mcParticle.pdgCode()) == Pdg::kDS) {
-        auto mcCollision = mcParticle.template mcCollision_as<soa::Join<aod::McCollisions, aod::MultsExtraMC>>();
-        multiplicity = mcCollision.multMCFT0A() + mcCollision.multMCFT0C(); // multFT0M = multFt0A + multFT0C
-        hCandidates->Fill(kCandidateStepMcGenAll, mcParticle.pt(), multiplicity, mcParticle.originMcGen());
-        if (std::abs(mcParticle.flagMcMatchGen()) == 1 << aod::hf_cand_3prong::DecayType::DsToKKPi) {
-          hCandidates->Fill(kCandidateStepMcGenDsToKKPi, mcParticle.pt(), multiplicity, mcParticle.originMcGen());
-          auto yDs = RecoDecay::y(mcParticle.pVector(), o2::constants::physics::MassDS);
-          if (std::abs(yDs) <= yCandGenMax) {
-            hCandidates->Fill(kCandidateStepMcCandInAcceptance, mcParticle.pt(), multiplicity, mcParticle.originMcGen());
-          }
-          bool isDaughterInAcceptance = true;
-          auto daughters = mcParticle.template daughters_as<CandDsMcGen>();
-          for (const auto& daughter : daughters) {
-            if (daughter.pt() < ptDaughterMin || std::abs(daughter.eta()) > etaTrackMax) {
-              isDaughterInAcceptance = false;
-            }
-          }
-          if (isDaughterInAcceptance) {
-            hCandidates->Fill(kCandidateStepMcDaughtersInAcceptance, mcParticle.pt(), multiplicity, mcParticle.originMcGen());
-            fillHistoMcGen(mcParticle);
-          }
-        }
-      }
-    }
-
-    // recontructed candidates loop
-    for (auto& candidate : candidates) {
-      auto collision = candidate.template collision_as<soa::Join<aod::Collisions, aod::FT0Mults>>();
-      multiplicity = collision.multFT0M();
-      registry.fill(HIST("hMultFT0M"), multiplicity);
-      if (std::abs(candidate.flagMcMatchRec()) == 1 << aod::hf_cand_3prong::DecayType::DsToKKPi) {
-        auto prong0McPart = candidate.template prong0_as<aod::TracksWMc>().template mcParticle_as<CandDsMcGen>();
-        // DsToKKPi and DsToPiKK division
-        if (((std::abs(prong0McPart.pdgCode()) == kKPlus) && (candidate.isSelDsToKKPi() >= selectionFlagDs)) || ((std::abs(prong0McPart.pdgCode()) == kPiPlus) && (candidate.isSelDsToPiKK() >= selectionFlagDs))) {
-          registry.fill(HIST("hPtCand"), candidate.pt());
-          hCandidates->Fill(kCandidateStepMcReco, candidate.pt(), multiplicity, candidate.originMcRec());
-          if (std::abs(hfHelper.yDs(candidate)) <= yCandMax) {
-            hCandidates->Fill(kCandidateStepMcRecoInAcceptance, candidate.pt(), multiplicity, candidate.originMcRec());
-          }
-        }
-      }
-    }
-  }
-  PROCESS_SWITCH(HfCorrelatorDsHadrons, processMcCandEfficiency, "Process MC for calculating candidate reconstruction efficiency", false);
-
-  /// Ds-Hadron correlation - for calculating associated particle tracking efficiency using MC reco-level analysis
-  void processMcTrackEfficiency(soa::Join<aod::Collisions, aod::FT0Mults, aod::EvSels> const&,
-                                soa::Join<aod::McCollisions, aod::MultsExtraMC> const&,
-                                CandDsMcGen const& mcParticles,
-                                TracksWithMc const& tracksData)
-  {
-    auto hAssocTracks = registry.get<StepTHn>(HIST("hAssocTracks"));
-
-    /// Gen loop
-    float multiplicity = -1.;
-    float posZ = -20.;
-    for (auto& mcParticle : mcParticles) {
-      // generated tracks
-      if (mcParticle.isPhysicalPrimary() && ((std::abs(mcParticle.pdgCode()) == kElectron) || (std::abs(mcParticle.pdgCode()) == kMuonMinus) || (std::abs(mcParticle.pdgCode()) == kPiPlus) || (std::abs(mcParticle.pdgCode()) == kKPlus) || (std::abs(mcParticle.pdgCode()) == kProton))) {
-        auto mcCollision = mcParticle.template mcCollision_as<soa::Join<aod::McCollisions, aod::MultsExtraMC>>();
-        multiplicity = mcCollision.multMCFT0A() + mcCollision.multMCFT0C(); // multFT0M = multFt0A + multFT0C
-        posZ = mcCollision.posZ();
-        hAssocTracks->Fill(kAssocTrackStepMcGen, mcParticle.eta(), mcParticle.pt(), multiplicity, posZ);
-        if (mcParticle.pt() > ptTrackMin && std::abs(mcParticle.eta()) < etaTrackMax) {
-          hAssocTracks->Fill(kAssocTrackStepMcGenInAcceptance, mcParticle.eta(), mcParticle.pt(), multiplicity, posZ);
-          registry.fill(HIST("hPtParticleAssocMcGen"), mcParticle.pt());
-        }
-      }
-    }
-
-    // recontructed tracks loop
-    for (auto& track : tracksData) {
-      if (track.has_collision()) {
-        if (!track.isGlobalTrackWoDCA() || track.dcaXY() > dcaXYTrackMax || track.dcaZ() > dcaZTrackMax || track.tpcNClsCrossedRows() < nTpcCrossedRaws) {
-          continue;
-        }
-        auto collision = track.template collision_as<soa::Join<aod::Collisions, aod::FT0Mults, aod::EvSels>>();
-        if (useSel8ForTrackEff && !collision.sel8()) {
-          continue;
-        }
-        multiplicity = collision.multFT0M();
-        posZ = collision.posZ();
-        registry.fill(HIST("hZVtx"), posZ);
-        registry.fill(HIST("hMultFT0M"), multiplicity);
-        hAssocTracks->Fill(kAssocTrackStepRecoAll, track.eta(), track.pt(), multiplicity, posZ);
-        if (track.has_mcParticle()) {
-          auto mcParticle = track.template mcParticle_as<CandDsMcGen>();
-          auto mcCollision = mcParticle.template mcCollision_as<soa::Join<aod::McCollisions, aod::MultsExtraMC>>();
-          if (eventGeneratorType >= 0 && mcCollision.getSubGeneratorId() != eventGeneratorType) {
-            continue;
-          }
-          hAssocTracks->Fill(kAssocTrackStepRecoMcMatch, mcParticle.eta(), mcParticle.pt(), multiplicity, posZ);
-          if (mcParticle.isPhysicalPrimary()) {
-            hAssocTracks->Fill(kAssocTrackStepRecoPrimaries, mcParticle.eta(), mcParticle.pt(), multiplicity, posZ);
-            registry.fill(HIST("hPtParticleAssocMcRec"), track.pt());
-            if ((std::abs(mcParticle.pdgCode()) == kElectron) || (std::abs(mcParticle.pdgCode()) == kMuonMinus) || (std::abs(mcParticle.pdgCode()) == kPiPlus) || (std::abs(mcParticle.pdgCode()) == kKPlus) || (std::abs(mcParticle.pdgCode()) == kProton)) {
-              hAssocTracks->Fill(kAssocTrackStepRecoSpecies, mcParticle.eta(), mcParticle.pt(), multiplicity, posZ);
-              if (mcParticle.mcCollisionId() == track.collisionId()) {
-                hAssocTracks->Fill(kAssocTrackStepRecoCollMatch, mcParticle.eta(), mcParticle.pt(), multiplicity, posZ);
-              }
-            }
-          }
-        }
-      } else {
-        // fake track
-        hAssocTracks->Fill(kAssocTrackStepFake, track.eta(), track.pt(), multiplicity, posZ);
-      }
-    }
-  }
-  PROCESS_SWITCH(HfCorrelatorDsHadrons, processMcTrackEfficiency, "Process MC for calculating associated particle tracking efficiency", false);
 
   /// Ds-Hadron correlation pair builder - for MC gen-level analysis (no filter/selection, only true signal)
   void processMcGen(SelCollisionsWithDsMc::iterator const& mcCollision,
